refactor: 요청 응답 DTO application으로 옮김
