+++ menu/src/main/java/kitchenpos/menu/application/response/MenuProductResponse.java
@@ -1,0 +1,73 @@
+package kitchenpos.menu.application.response;
+
+import kitchenpos.menu.domain.MenuProduct;
+
+import java.util.Objects;
+
+public class MenuProductResponse {
+    private final Long seq;
+    private final Long menuId;
+    private final Long productId;
+    private final long quantity;
+
+    public MenuProductResponse(final Long seq,
+                               final Long menuId,
+                               final Long productId,
+                               final long quantity) {
+        this.seq = seq;
+        this.menuId = menuId;
+        this.productId = productId;
+        this.quantity = quantity;
+    }
+
+    public static MenuProductResponse from(final MenuProduct menuProduct) {
+        return new MenuProductResponse(
+                menuProduct.getProductId(),
+                menuProduct.getMenu().getId(),
+                menuProduct.getProductId(),
+                menuProduct.getQuantity()
+        );
+    }
+
+    public Long getSeq() {
+        return seq;
+    }
+
+    public Long getMenuId() {
+        return menuId;
+    }
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public long getQuantity() {
+        return quantity;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        final MenuProductResponse that = (MenuProductResponse) o;
+        return quantity == that.quantity
+                && Objects.equals(seq, that.seq)
+                && Objects.equals(menuId, that.menuId)
+                && Objects.equals(productId, that.productId);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(seq, menuId, productId, quantity);
+    }
+
+    @Override
+    public String toString() {
+        return "MenuProductResponse{" +
+                "seq=" + seq +
+                ", menuId=" + menuId +
+                ", productId=" + productId +
+                ", quantity=" + quantity +
+                '}';
+    }
+}
