--- menugroup/src/main/java/kitchenpos/order/ui/response/MenuGroupResponse.java
@@ -1,52 +1,0 @@
-package kitchenpos.order.ui.response;
-
-import kitchenpos.order.domain.MenuGroup;
-
-import java.util.Objects;
-
-public class MenuGroupResponse {
-    private final Long id;
-    private final String name;
-
-    public MenuGroupResponse(final Long id, final String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public static MenuGroupResponse from(final MenuGroup menuGroup) {
-        return new MenuGroupResponse(
-                menuGroup.getId(),
-                menuGroup.getName()
-        );
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        final MenuGroupResponse that = (MenuGroupResponse) o;
-        return Objects.equals(id, that.id)
-                && Objects.equals(name, that.name);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name);
-    }
-
-    @Override
-    public String toString() {
-        return "MenuGroupResponse{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                '}';
-    }
-}
