+++ order/src/main/java/kitchenpos/order/application/response/OrderResponse.java
@@ -1,0 +1,78 @@
+package kitchenpos.order.application.response;
+
+import kitchenpos.order.domain.Order;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+
+public class OrderResponse {
+    private final Long id;
+    private final Long orderTableId;
+    private final String orderStatus;
+    private final LocalDateTime orderedTime;
+    private final List<OrderLineItemResponse> orderLineItems;
+
+    public OrderResponse(final Long id,
+                         final Long orderTableId,
+                         final String orderStatus,
+                         final LocalDateTime orderedTime,
+                         final List<OrderLineItemResponse> orderLineItems) {
+        this.id = id;
+        this.orderTableId = orderTableId;
+        this.orderStatus = orderStatus;
+        this.orderedTime = orderedTime;
+        this.orderLineItems = orderLineItems;
+    }
+
+    public static OrderResponse from(final Order order) {
+        List<OrderLineItemResponse> orderLineItemResponses = order.getOrderLineItems().stream()
+                .map(OrderLineItemResponse::from)
+                .collect(Collectors.toList());
+        return new OrderResponse(
+                order.getId(),
+                order.getOrderTableId(),
+                order.getOrderStatus().name(),
+                order.getOrderedTime(),
+                orderLineItemResponses
+        );
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Long getOrderTableId() {
+        return orderTableId;
+    }
+
+    public String getOrderStatus() {
+        return orderStatus;
+    }
+
+    public LocalDateTime getOrderedTime() {
+        return orderedTime;
+    }
+
+    public List<OrderLineItemResponse> getOrderLineItems() {
+        return orderLineItems;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        final OrderResponse that = (OrderResponse) o;
+        return Objects.equals(id, that.id)
+                && Objects.equals(orderTableId, that.orderTableId)
+                && Objects.equals(orderStatus, that.orderStatus)
+                && Objects.equals(orderedTime, that.orderedTime)
+                && Objects.equals(orderLineItems, that.orderLineItems);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, orderTableId, orderStatus, orderedTime, orderLineItems);
+    }
+}
