+++ ordertable/src/main/java/kitchenpos/ordertable/application/response/OrderTableResponse.java
@@ -1,0 +1,67 @@
+package kitchenpos.ordertable.application.response;
+
+import kitchenpos.ordertable.domain.OrderTable;
+
+import java.util.Objects;
+
+public class OrderTableResponse {
+    private final Long id;
+    private final Long tableGroupId;
+    private final int numberOfGuests;
+    private final boolean empty;
+
+    private OrderTableResponse(final Long id, final Long tableGroupId, final int numberOfGuests, final boolean empty) {
+        this.id = id;
+        this.tableGroupId = tableGroupId;
+        this.numberOfGuests = numberOfGuests;
+        this.empty = empty;
+    }
+
+    public static OrderTableResponse from(final OrderTable orderTable) {
+        return new OrderTableResponse(
+                orderTable.getId(),
+                orderTable.getTableGroupId() != null ? orderTable.getTableGroupId() : null,
+                orderTable.getNumberOfGuests(),
+                orderTable.isEmpty()
+        );
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Long getTableGroupId() {
+        return tableGroupId;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        final OrderTableResponse that = (OrderTableResponse) o;
+        return numberOfGuests == that.numberOfGuests && empty == that.empty && Objects.equals(id, that.id) && Objects.equals(tableGroupId, that.tableGroupId);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, tableGroupId, numberOfGuests, empty);
+    }
+
+    @Override
+    public String toString() {
+        return "OrdreTableResponse{" +
+                "id=" + id +
+                ", tableGroupId=" + tableGroupId +
+                ", numberOfGuests=" + numberOfGuests +
+                ", empty=" + empty +
+                '}';
+    }
+}
