+++ product/src/main/java/kitchenpos/product/application/response/ProductResponse.java
@@ -1,0 +1,58 @@
+package kitchenpos.product.application.response;
+
+import kitchenpos.product.domain.Product;
+
+import java.math.BigDecimal;
+import java.util.Objects;
+
+public class ProductResponse {
+    private final Long id;
+    private final String name;
+    private final BigDecimal price;
+
+    public ProductResponse(final Long id, final String name, final BigDecimal price) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+    }
+
+    public static ProductResponse from(final Product product) {
+        return new ProductResponse(product.getId(), product.getName(), product.getPrice());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        final ProductResponse that = (ProductResponse) o;
+        return Objects.equals(id, that.id)
+                && Objects.equals(name, that.name)
+                && Objects.equals(price, that.price);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name, price);
+    }
+
+    @Override
+    public String toString() {
+        return "ProductResponse{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", price=" + price +
+                '}';
+    }
+}
