+++ product/src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,8 +1,8 @@
 package kitchenpos.product.ui;
 
 import kitchenpos.product.application.ProductService;
-import kitchenpos.product.ui.request.ProductRequest;
-import kitchenpos.product.ui.response.ProductResponse;
+import kitchenpos.product.application.request.ProductRequest;
+import kitchenpos.product.application.response.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
