+++ web/src/test/java/kitchenpos/application/MenuGroupServiceTest.java
@@ -2,8 +2,8 @@ package kitchenpos.application;
 
 import kitchenpos.order.application.MenuGroupService;
 import kitchenpos.order.domain.repository.MenuGroupRepository;
-import kitchenpos.order.ui.request.MenuGroupRequest;
-import kitchenpos.order.ui.response.MenuGroupResponse;
+import kitchenpos.order.application.request.MenuGroupRequest;
+import kitchenpos.order.application.response.MenuGroupResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
