+++ web/src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -4,10 +4,10 @@ import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.repository.OrderRepository;
 import kitchenpos.order.domain.validator.OrderValidator;
-import kitchenpos.order.ui.request.OrderLineItemRequest;
-import kitchenpos.order.ui.request.OrderRequest;
-import kitchenpos.order.ui.request.UpdateOrderStatusRequest;
-import kitchenpos.order.ui.response.OrderResponse;
+import kitchenpos.order.application.request.OrderLineItemRequest;
+import kitchenpos.order.application.request.OrderRequest;
+import kitchenpos.order.application.request.UpdateOrderStatusRequest;
+import kitchenpos.order.application.response.OrderResponse;
 import kitchenpos.ordertable.domain.repository.OrderTableRepository;
 import kitchenpos.tablegroup.domain.repository.TableGroupRepository;
 import org.junit.jupiter.api.Test;
