+++ web/src/test/java/kitchenpos/application/TableServiceTest.java
@@ -3,10 +3,10 @@ package kitchenpos.application;
 import kitchenpos.order.domain.repository.OrderRepository;
 import kitchenpos.ordertable.application.TableService;
 import kitchenpos.ordertable.domain.repository.OrderTableRepository;
-import kitchenpos.ordertable.ui.request.OrderTableRequest;
-import kitchenpos.ordertable.ui.request.UpdateOrderTableEmptyRequest;
-import kitchenpos.ordertable.ui.request.UpdateOrderTableNumberOfGuestsRequest;
-import kitchenpos.ordertable.ui.response.OrderTableResponse;
+import kitchenpos.ordertable.application.request.OrderTableRequest;
+import kitchenpos.ordertable.application.request.UpdateOrderTableEmptyRequest;
+import kitchenpos.ordertable.application.request.UpdateOrderTableNumberOfGuestsRequest;
+import kitchenpos.ordertable.application.response.OrderTableResponse;
 import kitchenpos.tablegroup.domain.repository.TableGroupRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
